+++ tests/testthat/test-docx_exporter_functions.R
@@ -511,7 +511,7 @@ testthat::test_that("add_little_gap_bottom_borders_spanning_headers() works corr
 })
 
 
-testthat::test_that("export_as_docx_j() works with pagination", {
+testthat::test_that("export_TLG_as_docx() works with pagination", {
   # create a TableTree with a few pages
   colspan_trt_map <- data.frame(
     colspan_trt = c("Active Study Agent", "Active Study Agent", " "),
@@ -614,6 +614,46 @@ testthat::test_that("export_as_docx_j() works with pagination", {
   # export it as docx
   output_dir <- tempdir()
   options(docx.add_datetime = FALSE)
+  export_TLG_as_docx(
+    result,
+    output_dir = output_dir,
+    orientation = "landscape",
+    tblid = "test1234",
+    nosplitin = list(cols = c(.trtvar, "rrisk_header")),
+    paginate = TRUE,
+    add_page_break = TRUE,
+    combined_docx = TRUE,
+    export_csv = TRUE
+  )
+  options(docx.add_datetime = TRUE)
+
+  # check that the files exist, including the allparts
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234part1of2.docx")))
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234part2of2.docx")))
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234allparts.docx")))
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234part1of2.csv")))
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234part2of2.csv")))
+
+  # open the files and check the XML
+  doc <- officer::read_docx(paste0(output_dir, "/test1234part1of2.docx"))
+  snapshot_test_docx(doc)
+
+  doc <- officer::read_docx(paste0(output_dir, "/test1234part2of2.docx"))
+  snapshot_test_docx(doc)
+
+  doc <- officer::read_docx(paste0(output_dir, "/test1234allparts.docx"))
+  snapshot_test_docx(doc)
+
+  file.remove(c(
+    paste0(output_dir, "/test1234part1of2.docx"),
+    paste0(output_dir, "/test1234part2of2.docx"),
+    paste0(output_dir, "/test1234allparts.docx"),
+    paste0(output_dir, "/test1234part1of2.csv"),
+    paste0(output_dir, "/test1234part2of2.csv")
+  ))
+
+  # do not save csv
+  options(docx.add_datetime = FALSE)
   export_as_docx_j(
     result,
     output_dir = output_dir,
@@ -630,26 +670,55 @@ testthat::test_that("export_as_docx_j() works with pagination", {
   testthat::expect_true(file.exists(paste0(output_dir, "/test1234part1of2.docx")))
   testthat::expect_true(file.exists(paste0(output_dir, "/test1234part2of2.docx")))
   testthat::expect_true(file.exists(paste0(output_dir, "/test1234allparts.docx")))
+  testthat::expect_false(file.exists(paste0(output_dir, "/test1234part1of2.csv")))
+  testthat::expect_false(file.exists(paste0(output_dir, "/test1234part2of2.csv")))
+  file.remove(c(
+    paste0(output_dir, "/test1234part1of2.docx"),
+    paste0(output_dir, "/test1234part2of2.docx"),
+    paste0(output_dir, "/test1234allparts.docx")
+  ))
 
-  # open the files and check the XML
-  doc <- officer::read_docx(paste0(output_dir, "/test1234part1of2.docx"))
-  snapshot_test_docx(doc)
+  # save csv in a different location
+  output_csv_directory <- tempfile()
+  dir.create(output_csv_directory, showWarnings = FALSE, recursive = TRUE)
+  options(docx.add_datetime = FALSE)
+  printed_messages <- testthat::capture_messages(
+    export_as_docx_j(result,
+                     output_dir = output_dir,
+                     orientation = "landscape",
+                     tblid = "test1234",
+                     nosplitin = list(cols = c(.trtvar, "rrisk_header")),
+                     paginate = TRUE,
+                     add_page_break = TRUE,
+                     combined_docx = TRUE,
+                     export_csv = TRUE,
+                     output_csv_directory = output_csv_directory)
+  )
+  options(docx.add_datetime = TRUE)
 
-  doc <- officer::read_docx(paste0(output_dir, "/test1234part2of2.docx"))
-  snapshot_test_docx(doc)
-
-  doc <- officer::read_docx(paste0(output_dir, "/test1234allparts.docx"))
-  snapshot_test_docx(doc)
+  # check that the files exist, including the allparts
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234part1of2.docx")))
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234part2of2.docx")))
+  testthat::expect_true(file.exists(paste0(output_dir, "/test1234allparts.docx")))
+  testthat::expect_false(file.exists(paste0(output_dir, "/test1234part1of2.csv")))
+  testthat::expect_false(file.exists(paste0(output_dir, "/test1234part2of2.csv")))
+  testthat::expect_true(file.exists(paste0(output_csv_directory, "/test1234part1of2.csv")))
+  testthat::expect_true(file.exists(paste0(output_csv_directory, "/test1234part2of2.csv")))
+  testthat::expect_true(all.equal(printed_messages,
+    c(paste0("Saving csv as ", output_csv_directory, "/test1234part1of2.csv\n"),
+      paste0("Saving csv as ", output_csv_directory, "/test1234part2of2.csv\n"))
+  ))
 
   file.remove(c(
     paste0(output_dir, "/test1234part1of2.docx"),
     paste0(output_dir, "/test1234part2of2.docx"),
     paste0(output_dir, "/test1234allparts.docx")
   ))
+  unlink(output_csv_directory, recursive = TRUE)
 })
 
 
-testthat::test_that("export_graph_as_docx() works with basic example", {
+testthat::test_that("export_TLG_as_docx() works with basic example", {
   # create a few ggplots
   cbbPalette <- c("#000000", "#E69F00", "#0072B2")
 
@@ -767,7 +836,7 @@ testthat::test_that("export_graph_as_docx() works with basic example", {
   # export them as docx
   options(docx.add_datetime = FALSE)
   testthat::expect_no_error(
-    export_graph_as_docx(
+    export_TLG_as_docx(
       plotnames = list(pn1, pn2),
       tblid = "testgraph1234",
       output_dir = output_dir,