+++ tests/testthat/test-tt_to_tblfile.R
@@ -196,7 +196,7 @@ test_that("tt_to_tlgrtf works with wide table", {
     analyze("AGE")
 
   tbl_wide <- build_table(lyt_wide, ex_adsl)
-  expect_silent(suppressMessages(res_wide <- rtf_out_wrapper(tbl_wide, "test2", part = NA)))
+  expect_silent(suppressMessages(res_wide <- rtf_out_wrapper(tbl_wide, "test2", part = NA, export_csv = TRUE)))
   for (fl in res_wide) {
     expect_snapshot_file(compare = compare_file_text, fl, cran = TRUE)
     expect_snapshot_file(compare = compare_file_text, gsub("rtf$", "csv", fl))