+++ CHANGELOG.md
@@ -5,28 +5,38 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.4] - Unreleased
+## [0.1.5] - Unreleased
 
 ### Fixed
 - Fixed `s_kaplan_meier()` range censoring indicator handling to no longer produce `NA` values in the output when either all subjects are censored or none are censored.
 - Fixed the hanging indent in the first column of the body of the table (#138)
 - Export `leftside()`, `postfun_eq5d` `ac_blank_line` and `tt_to_tblfile`
 - Minor bugfix in DOCX exporter when calculating the number of pages (#188)
+- Fixed `tt_to_flextable_j()` to have correct left-indentation in header col 1 (#171) and footer first line starting with newline (#171)
 
 ### Changed
 - refactored functions `tt_to_flextable_j()` and `export_as_docx_j()`
+- created generic wrapper function `export_TLG_as_docx()`, which now calls `export_as_docx_j()` and `export_graph_as_docx()` (#173)
+- Functions `export_as_docx_j()` and `export_graph_as_docx()` still exist but are not internal, i.e. not exported
 - updated vignette to explain correctly how to insert newlines in the headers of Tables and Listings (#179)
 
-## [0.1.3.9] - Unreleased
+## [0.1.4] - 2026-02-05
+
+- [0.1.3.9.1] and [0.1.3.9.2] merged together
+
+## [0.1.3.9.1] - 2026-02-02
 
 ### Fixed
 - Fixed `tt_to_tlgrtf()` argument `label_width_ins` which was not applying the change in the row label column width (#166).
+
+### Added
+- Added option to switch on/off the export of the csv in both `tt_to_tlgrtf()` and `export_as_docx_j()`
+- Added option to specify the output folder for the csv
 
 ## [0.1.3] - 2026-01-12
 
 ### Changed
 - Address CRAN NOTES.
-
 
 ## [0.1.2] - 2025-12-10